Moved certain tests to seperate file and excluded from the FF suite due to incompability with selenium FF driver
