+++ selenium/config/wdio.conf.js
@@ -15,7 +15,6 @@ let config = {
   // directory is where your package.json resides, so `wdio` will be called from there.
   //
   specs: [
-    // Patterns to exclude.
     './selenium/**/*.test.js'
   ],
   exclude: [
