+++ selenium/tests/borchk.test.js
@@ -57,93 +57,6 @@ describe('Test Borchk component', () => {
     assert.equal(hidden, 2);
   });
 
-  it('should select current selected element, close dropdown and focus next input field when tab is pressed', () => {
-    browser.addValue('#libraryid-input', 'rin');
-
-    // Pressing ArrowDown key -- https://w3c.github.io/webdriver/webdriver-spec.html#h-keyboard-actions
-    browser.keys('ArrowDown');
-
-    // Get text of currently selected item in dropdrown
-    const selectedElementText = browser.element('.selected').getText();
-
-    // Pressing Tab key
-    browser.keys('Tab');
-
-    // Get value of library input field
-    const inputValue = browser.element('#libraryid-input').getValue();
-
-    assert.equal(selectedElementText, inputValue);
-
-    // Assert that the userid input field have focus
-    assert.isTrue(browser.hasFocus('#userid-input'));
-  });
-
-  it('should select current selected element, close dropdown and focus next input field when enter is pressed', () => {
-    browser.addValue('#libraryid-input', 'rin');
-
-    // Pressing ArrowDown key -- https://w3c.github.io/webdriver/webdriver-spec.html#h-keyboard-actions
-    browser.keys('ArrowDown');
-
-    // Get text of currently selected item in dropdrown
-    const selectedElementText = browser.element('.selected').getText();
-
-    // Pressing Enter key
-    browser.keys('Enter');
-
-    // Get value of library input field
-    const inputValue = browser.element('#libraryid-input').getValue();
-
-    assert.equal(selectedElementText, inputValue);
-
-    // Assert that the userid input field have focus
-    assert.isTrue(browser.hasFocus('#userid-input'));
-  });
-
-  it('should close dropwdown on escape', () => {
-    assert.isFalse(browser.isVisible('#libraries-dropdown'));
-    browser.addValue('#libraryid-input', 'rin');
-    assert.isTrue(browser.isVisible('#libraries-dropdown'));
-    // Pressing escape key -- https://w3c.github.io/webdriver/webdriver-spec.html#h-keyboard-actions
-    browser.keys('Escape');
-    assert.isFalse(browser.isVisible('#libraries-dropdown'));
-  });
-
-  it('should add selected items to "Forslag"', () => {
-    // Empty localStorage
-    browser.localStorage('DELETE', 'agencies');
-    // Refresh browser to make it take effect
-    browser.refresh();
-
-    browser.addValue('#libraryid-input', 'rin');
-
-    // Ensure the two labels in the dropdown are hidden
-    assert.isFalse(browser.isVisible('#latest'));
-    assert.isFalse(browser.isVisible('#alphabetical'));
-
-
-    // Select first item in dropdown -- ArrowDown Enter
-    browser.keys(['ArrowDown', 'Enter']);
-
-    // We should now have one item with the .recent class set
-    assert.equal(browser.elements('.recent').value.length, 1);
-
-    browser.setValue('#libraryid-input', '');
-    browser.addValue('#libraryid-input', 'rin');
-
-    // Ensure the two labels in the dropdown are now visible
-    assert.isTrue(browser.isVisible('#latest'));
-    assert.isTrue(browser.isVisible('#alphabetical'));
-
-    // ArrowDown ArrowDown ArrowDown Enter
-    browser.keys(['ArrowDown', 'ArrowDown', 'ArrowDown', 'Enter']);
-
-    browser.setValue('#libraryid-input', '');
-    browser.addValue('#libraryid-input', 'rin');
-
-    // We should now have one item with the .recent class set
-    assert.equal(browser.elements('.recent').value.length, 2);
-  });
-
   it('Should toggle the libraries button open/closed', () => {
     // libraries dropdown should be invisible
     assert.isFalse(browser.isVisible('#libraries-dropdown'));
